Finish Authentication & Apply bootstrap

+++ articles/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST, require_GET
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
@@ -25,9 +27,8 @@ def detail(request, article_pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 def create(request):
-    # if request.method == 'POST':
-    #     # Article 을 생성해달라고 하는 요청
     form = ArticleForm(request.POST or None)
     if form.is_valid():
         article = form.save()
@@ -41,6 +42,7 @@ def create(request):
     return render(request, 'articles/create.html', context)
 
 
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     form = ArticleForm(request.POST or None, instance=article)
@@ -53,28 +55,33 @@ def update(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
 
+@require_POST
 def create_comment(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    form = CommentForm(request.POST or None)
-    if form.is_valid():
-        comment = form.save(commit=False)
-        comment.article = article
-        comment.save()
+    if request.user.is_authenticated:
+        form = CommentForm(request.POST or None)
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.article = article
+            comment.save()
     return redirect('articles:detail', article_pk)
     # context = {
     #     'article': article,
     #     'form': form,
     # }
     # return render(request, 'articles/detail.html', context)
-    
+
 
 @require_POST
-def delete_comment(require, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+def delete_comment(request, article_pk, comment_pk):
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    return HttpResponse('Your are Unauthorized', status=401)  # 401 -> 인증되지 않았다
